Minor changes and testing MD_Parola
Testing MD Parola on an ESP32 S3 Mini
so far seems to be working fine with one panel.
Need level Shifters to drive more than one display since ESP is 3v3 Logic and the MAX7219 is 5V.
The low threshold is 3.7. the drop to a second panel is enough to not detect a logic high on 3v3..
Only using example code at the moment will be fleshing it out at some point. The rest of my logic can most likely be built on the Arduino and the use of the Functions can be wrapped since they just display text.
Willcontinue testing MD_Parola

Also Uploading some pin diagrams